+++ src/bpffeature.h
@@ -91,7 +91,6 @@ public:
   bool has_btf_func_global();
   bool has_map_batch();
   bool has_d_path();
-  bool has_uprobe_refcnt();
   bool has_kprobe_multi();
   bool has_kprobe_session();
   bool has_uprobe_multi();
@@ -133,7 +132,6 @@ protected:
   std::optional<bool> has_d_path_;
   std::optional<int> insns_limit_;
   std::optional<bool> has_map_batch_;
-  std::optional<bool> has_uprobe_refcnt_;
   std::optional<bool> has_kprobe_multi_;
   std::optional<bool> has_kprobe_session_;
   std::optional<bool> has_uprobe_multi_;